// src/app/modules/comercial/ciclo-vendas/autorizaciones/formulario/models/carrinho.ts
export interface ICarrinhoModel {
  idReservado: number;
  codCotacao: number;
  nrItem: number;
  codMaterial: number;
  nomeMaterial: string;
  codEmpresa: number;
  nomeEmpresa: string;
  codDeposito: number;
  nomeDeposito: string;
  quantidade: number;
  valorUnit: number;
  valor: number;
  percentualIpi: number;
  valorIpi: number;
  percentualIcms: number;
  valorIcms: number;
  valorIcmsSt: number;
  valorConcorrente?: number;
  tipoDesc: string;
  percentualDesc: number;
  valorDesc: number;
  valorTotalOri: number;
  valorTotal: number;
  medida1: number;
  medida2: number;
  pesoEspecifico: number;
  fatorMultiplo: number;
  materialAssociado: number;
  valorMaterialBarra:  number;
  valorMaterialContrato:number;
  valorMaterialPreco: number;
  valorServicoApsContrato:number;
  valorServicoContrato:number;
  qtdeItem: number;
  unidade: string;
  qtdePecas: number;
  valorBaseIcmsSt: number;
  aliquotaReducaoIcms: number;
  quantidadeItem: number;
  sequenciaLote: number;
  controladoPorLote: number;
  estoqueLote: number;
  loteFabricacao: string;
  nrPedidoCliente: string;
  codItemPedidoCliente: string;
  codProdutoCliente: string;
}